Mariabackup: Use fast path for srv_undo_tablespaces_init()
When applying the incremental backup, we need to initialize the
metadata for undo tablespaces (if any). When doing this, we must not
try to access the files, because that would require applying the
redo log, and we cannot do that before the .delta files have been
applied.

Because xtrabackup --prepare always skips any undo log processing,
it does not need to recover the transaction system state in any way.